third_party/blink/web_tests/platform/linux-chrome/external/wpt/webcodecs/videoFrame-construction.any-expected.txt

This is a testharness.js-based test.
[FAIL] Test we can construct a VideoFrame from a <video>.
  assert_equals: codedWidth expected 320 but got 384
Harness: the test ran to completion.
